* process.c: suppress warning for signed and unsigned type   inconsistency.
* ext/psych/parser.c: ditto.

* ext/sdbm/_sdbm.c: ditto.

* ext/syck/rubyext.c: ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@27864 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/sdbm/_sdbm.c
@@ -861,7 +861,7 @@ chkpage(char *pag)
 	register int off;
 	register short *ino = (short *) pag;
 
-	if ((n = GET_SHORT(ino,0)) < 0 || n > PBLKSIZ / sizeof(short))
+	if ((n = GET_SHORT(ino,0)) < 0 || n > PBLKSIZ / (int)sizeof(short))
 		return 0;
 
 	if (n > 0) {
